requirements.txt
fastapi
uvicorn
pyrogram
tgcrypto
python-dotenv
psycopg2-binary
python-multipart